+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.56] - 2024-12-19
+
+### Added
+- Company form: Logo upload functionality when editing a company
+- Person detail: Company logos now displayed in work history section instead of generic icon
+
 ## [1.0.55] - 2024-12-19
 
 ### Changed